+++ lib/shopify_api/config.ex
@@ -1,0 +1,6 @@
+defmodule ShopifyAPI.Config do
+  @moduledoc false
+
+  def lookup(key), do: Application.get_env(:shopify_api, key)
+  def lookup(key, subkey), do: Application.get_env(:shopify_api, key)[subkey]
+end