+++ test/test_helper.exs
@@ -1,3 +1,3 @@
 ExUnit.start()
-{:ok, _} = ShopifyAPI.CacheSupervisor.start_link([])
+{:ok, _} = ShopifyAPI.Supervisor.start_link([])
 Application.ensure_all_started(:bypass)